"Updated reply_to functionality in messaging methods"
Changed the 'reply_to_message_id' parameter to a more general 'reply_to' parameter across multiple message send methods. This allows either an id or a raw.types.InputReplyToMessage instance to be used for reply threading. Adjusted checks on 'reply_to' to handle this change. This was done to offer more flexibility when handling replies.