"Allow reply to story in reply_to"
Expanded functionality of the 'reply_to' parameter in various messaging methods to include 'raw.types.InputReplyToStory'. Prior to this, 'reply_to' only accepted an id or a 'raw.types.InputReplyToMessage' instance. This broadens the usage of the 'reply_to' parameter to allow for replies to stories in addition to messages, increasing flexibility for users.