Introduce authorized flag in 'Start' method
The 'Start' method in 'pyrogram.Client' has been refactored to add an 'authorize' boolean flag. This flag is useful to determine whether the session is authorized or not. In case the session is not authorized, the function will raise a RuntimeError. It has also modified the user validation by checking its instance before initializing the client.