Refactor TCP transport and update dependencies
The TCP transport was significantly refactored for improved organization and structure. The code handling proxies was simplified and made more straightforward via the usage of the python_socks library. The old way of handling proxies was greatly reduced and replaced by python_socks. This change necessitated an update of 'requirements.txt' to include python-socks version 2.4.4 for asyncio support.