Remove redundant class in type.txt
The class that was previously present in the type.txt file under the compiler API template has been taken out. This class was redundant, as it wasn't providing any functionality that was needed and wasn't used in the program. The cleaned-up code simplifies the type interface and prevents potential confusion caused by unnecessary components in the code.